fast_api_server/requirements.txt
requests==2.27.1
uvicorn==0.17.4
python-dotenv==0.19.2
GitPython==3.1.26
fastapi==0.73.0
opentelemetry-sdk==1.9.1
opentelemetry-instrumentation-fastapi==0.28.b1
opentelemetry-instrumentation-requests==0.28.b1
opentelemetry-instrumentation-logging==0.28.b1
opentelemetry-instrumentation-digma==0.5.5

